secure-api-gateway-ig-extensions/src/main/java/com/forgerock/sapi/gateway/common/rest/ContentTypeFormatterJson.java
/*
 * Copyright © 2020-2024 ForgeRock AS (obst@forgerock.com)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.forgerock.sapi.gateway.common.rest;

import java.util.Map;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;

/**
 * Formats a Map of String key value pairs into a json string
 */
public class ContentTypeFormatterJson implements ContentTypeFormatter{

    private static final ObjectMapper objectMapper;

    static{
        objectMapper = new ObjectMapper();
    }

    @Override
    public String getFormattedResponse(Map<String, String> fields) {
        if(fields == null){
            return "{}";
        }
        try {
            return objectMapper.writeValueAsString(fields);
        } catch (JsonProcessingException e) {
            throw new RuntimeException(e);
        }
    }
}
